add multithreaded file locator

+++ src/main/java/com/nw/itext/pdfchanger/fileprocessors/PDFRollBack.java
@@ -1,25 +1,33 @@
 package com.nw.itext.pdfchanger.fileprocessors;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.util.List;
+
+import com.nw.itext.pdfchanger.app.ConfigLoader;
 
 public class PDFRollBack implements FileProcessorIF {
 	private String currentFilePath;
-	private String suffix;
 
-	public PDFRollBack(String filePath, String suffix) {
+	public PDFRollBack(String filePath) {
 		this.currentFilePath = filePath;
-		this.suffix = suffix;
-
 	}
 
 	public boolean processFile() {
-		Path bkpFilePath = Paths.get(this.currentFilePath + this.suffix);
-		Path originalFilePath = Paths.get(this.currentFilePath);
+		
+		String fileName=currentFilePath.substring(currentFilePath.lastIndexOf("/")+1);
+		String bkpFolderPath= currentFilePath.substring(0,currentFilePath.lastIndexOf("/"))+ConfigLoader.getInstance().getBkpSuffix();
+		File bkpFolder=new File(bkpFolderPath);
+		if(!bkpFolder.exists()){
+			System.err.println("Backup folder not found for file :"+ currentFilePath);
+			return false;
+		}
+		Path originalFilePath=Paths.get(currentFilePath);
+		Path bkpFilePath=Paths.get(bkpFolderPath+fileName);
+		
 		System.out.println(originalFilePath);
 		try {
 			Files.copy(bkpFilePath, originalFilePath,
